refactor: createGatheringModal (#186)
